Serialize and deserialize timestamps as strings
Fields that are known to be timestamps or durations are serialized and
deserialized as strings.

This is done to avoid loss of precision in JavaScript 'number' type for
values above Number.MAX_SAFE_INTEGER (2^53 - 1).

The trace-server is expected to support receiving those timestamps as
strings.

The client can handle receiving timestamps as string or number.

The models are updated to use 'bigint' type instead of 'number' for
these fields. When receiving timestamps, the string or number value will
be converted to a BigInt.

Signed-off-by: Patrick Tasse <patrick.tasse@ericsson.com>